# routes/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Music Stram descritption",
        "version": "1.0.0",
        "title": "Music Stream",
        "termsOfService": "",
        "contact": {
            "email": "musicstream@gmail.com"
        }
    },
    "host": "",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "paths": {  
...
        "/editBooking": {
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Edit booking details",
                "description": "",
                "operationId": "editBooking",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "bookingId",
                        "description": "",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "formData",
                        "name": "artistId",
                        "description": "",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "formData",
                        "name": "place",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "description",
                        "description": "",
                        "type": "string"
                    },
                    {
                        "in": "formData",
                        "name": "date",
                        "description": "yyyy/mm/dd",
                        "required": true,
                        "type": "string",
                        "format": "date"
                    },
                    {
                        "in": "formData",
                        "name": "time",
                        "description": "hh:mm:ss",
                        "required": true,
                        "type": "string",
                        "format": "time"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated /Not updated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/bookNowEvent": {
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Send email",
                "description": "",
                "operationId": "bookNowEvent",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
               "parameters": [    
                   {
                    "in": "formData",
                    "name": "userId",
                    "description": "",
                    "required": true,
                    "type": "integer"
                },            
                {
                    "in": "formData",
                    "name": "artistId",
                    "description": "",
                    "required": true,
                    "type": "integer"
                },
                {
                    "in": "formData",
                    "name": "eventId",
                    "description": "",
                    "required": true,
                    "type": "integer"
                },
                {
                    "in": "formData",
                    "name": "name",
                    "description": "",
                    "required": true,
                    "type": "string"
                },
                {
                    "in": "formData",
                    "name": "email",
                    "description": "",
                    "required": true,
                    "type": "string"
                },
                {
                    "in": "formData",
                    "name": "description",
                    "description": "",
                    "required": true,
                    "type": "string"
                }
                
                
            ],
                "responses": {
                    "200": {
                        "description": "1- sucess/0- fail"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
...
}